+++ app/scripts/controllers/cuePoints.js
@@ -30,10 +30,10 @@ angular.module('myApp').controller('CuePointsCtrl',
                 tracks: [
                     {
                         src: "assets/subs/pale-blue-dot.vtt",
-                        kind: "subtitles",
+                        kind: "captions",
                         srclang: "en",
                         label: "English",
-                        default: ""
+                        default: "default"
                     }
                 ]
             }