+++ app/scripts/controllers/dash.js
@@ -54,10 +54,10 @@ angular.module('myApp').controller('DashCtrl',
                 tracks: [
                     {
                         src: "assets/subs/pale-blue-dot.vtt",
-                        kind: "subtitles",
+                        kind: "captions",
                         srclang: "en",
                         label: "English",
-                        default: ""
+                        default: "default"
                     }
                 ]
             }